% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/configuration.R
\name{build_ids}
\alias{build_ids}
\title{Building Position IDs}
\usage{
build_ids(N)
}
\arguments{
\item{N}{N}
}
\value{
list - $N, $loc_ids, $loc_bit_ids: location ID in bit format
}
\description{
Building position IDs
}
\examples{
N=4
K=2
data <- build_ids(N)
data$K = K
fun <- landscape_gen(N,K)
fitness_values <- foreach(id=t(as.matrix(data$loc_bit_ids)),.combine=c) \%do\% fun(as.numeric(id))
nk_landscape <- cbind(data$loc_bit_ids,fitness_values)
data$fitness_values <- fitness_values
data$nk_landscape <- nk_landscape
}

