% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{calculate_value_on_location}
\alias{calculate_value_on_location}
\title{Evaluating Possible Alternatives to Select a New Location ID}
\usage{
calculate_value_on_location(current_fit, nk.result, nk_landscape, noise = 0,
  noise_vector = NULL)
}
\arguments{
\item{current_fit}{Fitness value of current location}

\item{nk.result}{Alternative information}

\item{nk_landscape}{}

\item{noise}{Noise coefficient (0=none to 1=full)}

\item{noise_vector}{Noise structure (default, following a uniform distribution)}
}
\value{
result list
}
\description{
Evaluating Possible Alternatives to Select a New Location ID
}
\seealso{
\code{\link{search_hill_climbing}}
}

