% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nkm_r.R
\name{convert_influence_contribution}
\alias{convert_influence_contribution}
\title{Creating a Contribution Matrix from an Influence Matrix}
\usage{
convert_influence_contribution(inf_mat)
}
\arguments{
\item{inf_mat}{influence matrix (N by N)}
}
\value{
contribution matrix PI
}
\description{
Getting PI for generate an NK Landscape
}
\examples{
inf_mat <- matrix(c(1,1,1,0,0,1,1,1,1,0,1,1,1,1,0,1),4)
cont_mat <- convert_influence_contribution(inf_mat)
landscape_fun_N4_K2 = landscape_gen(N=4,K=2,PI=cont_mat)
landscape_fun_N4_K2(c(0,1,1,0))
}

