% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nkm_r.R
\name{create_landscape}
\alias{create_landscape}
\title{Creating an NK Landscape}
\usage{
create_landscape(N, K, fuc)
}
\arguments{
\item{N}{N}

\item{K}{K}

\item{fuc}{an object of landscape_gen or landscape_gen_lowdim}
}
\value{
a list of fitness values and others:
\enumerate{
\item $N
\item $K
\item $loc_ids
\item $loc_bit_ids
\item $landscape
}
}
\description{
Creating all fitness values based on a given generation function
}

