% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nkm_r.R
\name{create_locations}
\alias{create_locations}
\title{Creating Location IDs and Bitwise Representation}
\usage{
create_locations(N)
}
\arguments{
\item{N}{}
}
\value{
data.frame object of location ids and bitwise representation
}
\description{
Creating location IDs and bitwise representation
}

