% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/configuration.R
\name{gen_lowdim_fraction}
\alias{gen_lowdim_fraction}
\title{Generating Low Dimensionality Fractions}
\usage{
gen_lowdim_fraction(configure, N1)
}
\arguments{
\item{configure}{Original configuration}

\item{N1}{Low dimensionality masking vector}
}
\value{
Low dimensionality fractions (matrix)
}
\description{
Generating Low Dimensionality Fractions
}

