% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_all_configuration}
\alias{get_all_configuration}
\title{Looking around All Possibile Configuration Changes}
\usage{
get_all_configuration(location_id, N, coverage = 1:N)
}
\arguments{
\item{location_id}{Agent's location}

\item{N}{Number of components}

\item{coverage}{}
}
\value{
Location IDs for all the possible configurations
}
\description{
Looking around all the possible configuration changes
}

