% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_configuration}
\alias{get_configuration}
\title{Getting a New Location ID by Configuring a given Component}
\usage{
get_configuration(location_id, N, element_id)
}
\arguments{
\item{location_id}{Agent's location}

\item{N}{Number of components}

\item{element_id}{Component ID (from 1 to N)}
}
\value{
Location ID
}
\description{
Getting a New Location ID by Configuring a given Component
}
\seealso{
\code{\link{get_all_configuration}}
}

