% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nkm_r.R
\name{landscape_gen}
\alias{landscape_gen}
\title{Fitness Contribution Function Generator}
\usage{
landscape_gen(N = 4, K = 0, PI = 1:K, g = NULL)
}
\arguments{
\item{N}{Node number (bit code count)}

\item{K}{dependencies}

\item{PI}{determining neighbors, default is one-bit off...}

\item{g}{contribution matrix structure (N by 2^(K+1)), default follows a uniform distribution}
}
\value{
a function(x), x is a numeric vector with 0 or 1, and the length equals N
}
\description{
This function generates a fitness contribution function
}
\examples{
uniform_landscape_N4_K0 <- landscape_gen(N=4,K=0)
uniform_landscape_N4_K1 <- landscape_gen(4,1)
uniform_landscape_N4_K0(c(0,0,0,0))
uniform_landscape_N4_K0(c(1,0,0,0))
uniform_landscape_N4_K1(c(0,1,1,0))
}
\seealso{
CEGO::benchmarkGeneratorNKL
}

