% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nkm_r.R
\name{landscape_structure_uniform}
\alias{landscape_structure_uniform}
\title{Generating Random Numbers by Uniform Distribution}
\usage{
landscape_structure_uniform(N = 4, K = 0, seed_no = NULL)
}
\arguments{
\item{N}{}

\item{K}{}

\item{seed_no}{Random seed if any}
}
\value{
Uniform distribution
}
\description{
Generating Random Numbers by Uniform Distribution
}

