% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nkm_r.R
\name{redesign_influence_matrix}
\alias{redesign_influence_matrix}
\title{Redesign Influence Matrix with Keeping a Structure}
\usage{
redesign_influence_matrix(inf_mat, new_configuration)
}
\arguments{
\item{inf_mat}{Influence matrix}

\item{new_configuration}{New sequence of configuration}
}
\value{
a list of new influence matrix (a=original, b=transformed order)
}
\description{
Given a newly specified component structure, the influence matrix is redesigned.
}
\examples{
inf_mat <- matrix(c(1,1,1,0,0,1,1,1,1,0,1,1,1,1,0,1),4)
rownames(inf_mat) <- colnames(inf_mat) <- c("a","b","c","d")
new_configuration <- c("b","a","d","c")
new_inf_mat <- redesign_influence_matrix(inf_mat,new_configuration)$a
}
\seealso{
Ethiraj, S. K. and Levinthal, D. (2004). "Bounded Rationality and the Search for Organizational Architecture: An Evolutionary Perspective on the Design of Organizations and Their Evolvability," \emph{Administrative Science Quarterly}, \strong{49}, 404-437.
}

