% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search_strategy.R
\name{search_hill_climbing}
\alias{search_hill_climbing}
\title{Hill Climbing Search}
\usage{
search_hill_climbing(agent_size, loc_ids, nk_landscape, fileout = NULL)
}
\arguments{
\item{agent_size}{Number of agents}

\item{loc_ids}{Location ids (numeric vector)}

\item{nk_landscape}{fitness (matrix)}

\item{fileout}{output record file}
}
\description{
Executing hill climbing search
}
\examples{
nk.data = create_nkm_simple(6,3)
agent_size = 10
result = search_hill_climbing(agent_size,nk.data$loc_ids,nk.data$nk_landscape)
}

