% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search_strategy.R
\name{search_hill_climbing_full}
\alias{search_hill_climbing_full}
\title{Hill Climbing Search (Full Coverage)}
\usage{
search_hill_climbing_full(agent_size, loc_ids, nk_landscape, noise = 0,
  noise_vector = NULL)
}
\arguments{
\item{agent_size}{Number of agents}

\item{loc_ids}{Location ids (numeric vector)}

\item{nk_landscape}{fitness (matrix)}

\item{noise}{Noise coefficient (0=none to 1=full)}

\item{noise_vector}{Noise structure (default, following a uniform distribution)}
}
\description{
Executing hill climbing search
}
\examples{
nk.data = create_nkm_simple(6,3)
agent_size = 10
result = search_hill_climbing_full(agent_size,nk.data$loc_ids,nk.data$nk_landscape)
}
\seealso{
\code{\link{calculate_value_on_location}}
}

