% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.R
\name{test_nkm}
\alias{test_nkm}
\title{NK Lanscape simulation Test}
\usage{
test_nkm(N, K, loudly = TRUE)
}
\arguments{
\item{N}{N}

\item{K}{K}
}
\value{
optimal locations
}
\description{
A simple test unit for NKM
}

